Organizing the kernel
The way to have dependencies on other directories
for teyjus is to add the paths of these
dependencies by hand to the depend file and do not
use the one generated by the makefile (and not
delete it). Module imbed has such a configuration.

Also adding a toplevel makefile for making all
kernels at one.

git-svn-id: https://scm.gforge.inria.fr/svn/parsifal/code/proverif@9750 ae10dad5-531c-0410-a33e-f483eea19440